backend/requirements.txt
fastapi
uvicorn
asyncpg
python-multipart
httpx
pyjwt
cryptography
python-dotenv
